+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Improved data caching for debug link targets and DWP files
+- Improved kernel symbolization performance when no KASLR offset is
+  provided by the user
 
 
 0.2.1